+++ packages/cloud_firestore/CHANGELOG.md
@@ -1,7 +1,13 @@
-## 0.9.14
+## 0.10.1
 
 * Support for `startAtDocument`, `startAfterDocument`, `endAtDocument`, `endBeforeDocument`.
 * Added additional unit and integration tests.
+
+## 0.10.0
+
+* Support for `FieldValue.increment`.
+* Remove `FieldValue.type` and `FieldValue.value` from public API.
+* Additional integration testing.
 
 ## 0.9.13+1