+++ packages/webview_flutter/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.7
+
+* Fix loadUrlWithHeaders flaky test.
+
+## 0.3.6+1
+
+* Remove un-used method params in webview\_flutter
+
+## 0.3.6
+
+* Add an optional `headers` field to the controller.
+
 ## 0.3.5+5
 
 * Fixed error in documentation of `javascriptChannels`.